feat: 특정 면접 질문에 대한 모든 답변 조회 (#52)

+++ server/src/main/java/com/pullup/member/repository/CommentRepository.java
@@ -4,5 +4,5 @@ import com.pullup.interview.domain.Comment;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CommentRepository extends CrudRepository<Comment, Long> {
-    Long countByInterviewAnswerId(Long interviewAnswerId);
+    Integer countByInterviewAnswerId(Long interviewAnswerId);
 }
