+++ server/src/main/java/com/pullup/member/repository/LikeRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface LikeRepository extends CrudRepository<Like, Long> {
     Boolean existsByMemberIdAndInterviewAnswerId(Long memberId, Long interviewAnswerId);
-    Long countByInterviewAnswerId(Long interviewAnswerId);
+
+    Integer countByInterviewAnswerId(Long interviewAnswerId);
 }
